+++ utils/unique_vector_indices/unique_vector_indices.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-std::vector<int> unique_vector_indices(const std::vector<std::vector<double> > &raw_vectors,
-                                       const std::vector<std::vector<double> > &unique_vectors);
+std::vector<int> unique_vector_indices(const std::vector<std::vector<int> > &raw_vectors,
+                                       const std::vector<std::vector<int> > &unique_vectors);
 
 #endif // UNIQUE_VECTOR_INDICES_H
